+++ examples/global/src/app.js
@@ -19,7 +19,6 @@ render(
   <Provider store={store}>
     <div>
       <App />
-      <App />
       <hr />
       <OtherApp />
     </div>
