+++ lib/index.js
@@ -6,31 +6,10 @@ import config, { PAGE } from './config';
 import constants, { STORE, GET_VARIABLES, SET_VARIABLE, SET_VARIABLE_FN } from './constants';
 import reducers from './reducer';
 import reducerHelpers from './utils/reducer-helpers';
+import { setConfig, deleteConfig } from './utils/configs';
 import sagasArray, { saga as sagaObject } from './sagas';
 import ReSaga from './resaga';
 import selectors, { selectPage } from './selectors';
-
-const ResagaError = (error) => {
-  // TODO: will throw Error in the next minor release
-  // throw new Error(`[resaga] ${error}`);
-  console.error(`[resaga] Warning: ${error}. A breaking Error will be thrown in the next minor release.`); // eslint-disable-line no-console
-  return null;
-};
-
-const resagaConfigs = {};
-export const setConfig = (page, cfg) => {
-  if (resagaConfigs[page]) ResagaError(`Component '${page}' has been mounted multiple times which is anti-pattern. It's possibly due to dynamically generating by a loop. Consider moving state up to parent`);
-  resagaConfigs[page] = cfg;
-  return true;
-};
-export const getConfig = (page) => {
-  if (!page) return ResagaError(`The configuration of '${page}' not found. Check whether some component has been unmounted before the request finished, or it has never been mounted in the first place`);
-  const pageConfig = resagaConfigs[page];
-  if (!pageConfig) ResagaError(`Component '${page}' not found. Check whether the component has been unmounted before the request finished, or it has never been mounted in the first place`);
-
-  return pageConfig;
-};
-export const deleteConfig = (page) => delete resagaConfigs[page];
 
 export const reSagaHOC = (Component, configs) => {
   /**
