+++ lib/resaga.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { BEFORE_SUBMIT, SUBMIT, MANUALLY_CLEANUP } from './config';
 import { PAGE, GET_VARIABLES, SET_VARIABLE, SET_VARIABLE_FN, STORE, IS_LOADING, WILL_LOAD } from './constants';
 import component from './utils/component-helpers';
-import { getConfig } from './index';
+import { getConfig } from './utils/configs';
 
 export class ReSaga extends React.PureComponent {
   /**
