+++ lib/utils/component-helpers.js
@@ -1,6 +1,6 @@
 import { RESULT, SERVER_ERROR, STORE, SUBMIT_SUCCESS, IS_LOADING, WILL_LOAD, PAYLOAD } from '../constants';
 import { PAGE } from '../config';
-import { getConfig } from '../index';
+import { getConfig } from './configs';
 
 /**
  * @param request can be one of the 2 types below
